requirements.txt
discord.py>=2.5.0
yt-dlp>=2024.6.4
python-dotenv>=1.0.0
aiosqlite>=0.18.0
PyNaCl>=1.5.0
